Create Minimum stock level
In this code, we first create a dictionary called inventory that contains the current stock levels for each item in the system. We then create another dictionary called min_stock_levels that contains the minimum stock level for each item.

We then iterate over the items in the inventory dictionary using the items() method. For each item, we retrieve its minimum stock level from the min_stock_levels dictionary and compare it to its current stock level. If the current stock level is below the minimum stock level, we print an alert message that indicates which item's stock level is below the minimum level.