Create Items for Purchase order management
Add item to system: In this code, we define a class Item to represent the items in the system. The __init__ method initializes the attributes of the item with the provided name, description, and price.

The add_item function takes in the name, description, and price of the item to be added, as well as a list of existing items in the system. It creates a new instance of the Item class with the provided information, and adds it to the items list. Finally, it prints a message to confirm that the item has been added to the system.

In the example usage, we create an empty list items to hold the items, and call the add_item function to add a new item with the name "Apple", description "A juicy fruit", and price of 0.99.

Finding item using Unique Identifier: