Update Items for Purchase order management
Add Item to system: In this code, we define a class Item to represent the items in the system. The __init__ method initializes the attributes of the item with the provided name, description, and price.

The add_item function takes in the name, description, and price of the item to be added, as well as a list of existing items in the system. It creates a new instance of the Item class with the provided information, and adds it to the items list. Finally, it prints a message to confirm that the item has been added to the system.

In the example usage, we create an empty list items to hold the items, and call the add_item function to add a new item with the name "Apple", description "A juicy fruit", and price of 0.99.

Finding Item using Unique Identifier

Update Item information:  We define a dictionary called item with some initial information, including the item's name, price, and quantity. To update this item with new information, we simply assign new values to the corresponding keys in the dictionary. In this case, we update the price to 15.0 and the quantity to 10. Finally, we print the updated dictionary using the print function.

Deleting Item: In this code, we're using the os module which provides a way to interact with the operating system. We first check if the file exists by using the os.path.exists() function. If it exists, we delete the file using the os.remove() function and print a message confirming that the file has been deleted. If the file does not exist, we print a message indicating that.

Searching Item